% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{averageResponse}
\alias{averageResponse}
\title{Calculate the weighted average of estimated response values}
\usage{
averageResponse(weights, dose, modelResults, groupIndex = 1)
}
\arguments{
\item{weights}{numeric vector, estimated weights as returned by calculateWeights()}

\item{dose}{numeric vector, dose values at which response values are estimated}

\item{modelResults}{list, with results for each model, same length as weights. 
For each model a list with at least model.ans, regr.par, CES and ces.ans; 
these are by default included in result from f.proast(). Eventually 
contains also fct1 and fct2 if factors are included for the model parameters.}

\item{groupIndex}{integer, index for covariate combination of interest;
default value is 1; if NA average response for all groups is returned}
}
\value{
numeric (vector), averaged response value at given dose for group(s)
of interest
}
\description{
Calculate the weighted average of estimated response values
}
