% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{bindModelResults}
\alias{bindModelResults}
\title{Bind results of one/several fitted model(s) into one summary table}
\usage{
bindModelResults(savedResults)
}
\arguments{
\item{savedResults}{list as returned by the function fitAllModels() or 
fitAllCovariates()}
}
\value{
data frame containing for each model 
\itemize{
 \item{model: }{character, model name}
 \item{parameterNames: }{character, names of the parameters for which a covariate was included}
 \item{npar: }{integer, number of parameters in the model}
 \item{loglik: }{numeric, estimated log likelihood}
 \item{aic: }{numeric, estimated AIC}
 \item{bmd, bmdl, bmdu: }{per covariate value, the estimate, lower and upper
 bound of the benchmark dose}
 \item{converged: }{boolean, did fitting procedure ended with convergence}
}
}
\description{
Bind results of one/several fitted model(s) into one summary table
}
