% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{bmr1}
\alias{bmr1}
\title{Target function for optimization if ces.ans = 1}
\usage{
bmr1(weights, dose, modelResults, groupIndex = 1, baseResponse,
  maxResponse = NULL)
}
\arguments{
\item{weights}{numeric vector, estimated weights as returned by calculateWeights()}

\item{dose}{numeric, value for the dose at which response values are estimated}

\item{modelResults}{list, with results for each model, same length as weights. 
For each model a list with at least model.ans, regr.par, CES and ces.ans; 
these are by default included in result from f.proast(). Eventually 
contains also fct1 and fct2 if factors are included for the model parameters.}

\item{groupIndex}{integer, index for covariate combination of interest;
default value is 1; if NA results for all groups are returned}

\item{baseResponse}{numeric, estimated response value for dose = 0;
redundant but retained for convenience with bmr2() and bmr3()}

\item{maxResponse}{numeric, response value for which bmr should be calculated;
default value is NULL}
}
\value{
numeric (vector), estimated response value for a given dose and 
group(s) of interest; same object as returned by averageResponse()
}
\description{
Target function for optimization if ces.ans = 1
}
