% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{bmr3}
\alias{bmr3}
\title{Target function for optimization if ces.ans = 3}
\usage{
bmr3(weights, dose, modelResults, groupIndex = 1, baseResponse,
  maxResponse = NULL)
}
\arguments{
\item{weights}{numeric vector, estimated weights as returned by calculateWeights()}

\item{dose}{numeric, value for the dose at which response values are estimated}

\item{modelResults}{list, with results for each model, same length as weights. 
For each model a list with at least model.ans, regr.par, CES and ces.ans; 
these are by default included in result from f.proast(). Eventually 
contains also fct1 and fct2 if factors are included for the model parameters.}

\item{groupIndex}{integer, index for covariate combination of interest;
default value is 1; if missing average response for all groups is returned}

\item{baseResponse}{numeric, estimated response value for dose = 0}

\item{maxResponse}{numeric, response value for which bmr should be calculated;
default value is NULL}
}
\value{
numeric (vector), estimated extra risk for group(s) of interest
}
\description{
Target function for optimization if ces.ans = 3
}
