% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{calculateWeights}
\alias{calculateWeights}
\title{Calculate weights based on fitted models' aic}
\usage{
calculateWeights(aicValues)
}
\arguments{
\item{aicValues}{numeric vector, each model's estimated aic value}
}
\value{
numeric vector, estimated weight for each model.
Same order as aicValues but models with missing aicValues are excluded
}
\description{
Calculate weights based on fitted models' aic
}
