% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.basics.cat}
\alias{f.basics.cat}
\title{compute basic statistics of a categorical model}
\usage{
f.basics.cat(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{ans.all list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
update ans.all
}
\description{
compute basic statistics of a categorical model
}
