% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.bb.bin}
\alias{f.bb.bin}
\title{compute the value for parameter b, from values of other parameters}
\usage{
f.bb.bin(model.ans, aa, cc, dd, CED, CES, ces.ans, track = FALSE)
}
\arguments{
\item{model.ans}{integer, determines the type of model to be fitted}

\item{aa}{value for parameter a}

\item{cc}{value for parameter c}

\item{dd}{value for parameter d}

\item{CED}{numeric, value for the CED}

\item{CES}{numeric, value for the CES}

\item{ces.ans}{index type of benchmark response}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
value for the b parameter
}
\description{
From the original function from proast61.3, only the code for 
model.ans = [15, 16, 18, 19, 21, 24, 25, 26] is retained.
}
