% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.bb.con}
\alias{f.bb.con}
\title{Determine value for parameter b for continuous response}
\usage{
f.bb.con(model.ans, cc, dd, CED, CES, track = FALSE)
}
\arguments{
\item{model.ans}{integer, indicates the model that will be fitted;
one of \code{c(13, 15, 23, 25)}}

\item{cc}{numeric, value for parameter c}

\item{dd}{numeric, value for parameter d}

\item{CED}{numeric, value for the CED}

\item{CES}{numeric, value for the CES}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric, value for parameter b
}
\description{
Determine value for parameter b for continuous response
}
