% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.boot.con}
\alias{f.boot.con}
\title{One bootstrap run for continuous response}
\usage{
f.boot.con(ans.all.bt, track = FALSE)
}
\arguments{
\item{ans.all.bt}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, estimated MLEs, CED and whether the model converged
}
\description{
One bootstrap run for continuous response
}
