% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.cat}
\alias{f.cat}
\title{Main function for calculations with categorical response type}
\usage{
f.cat(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Main function for calculations with categorical response type
}
