% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.ced.cat}
\alias{f.ced.cat}
\title{Determine values for the CED; categorical model}
\usage{
f.ced.cat(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{ans.all list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list with matrix of CED, response matrix, pi background matrix and gr.txt
}
\description{
Determine values for the CED; categorical model
}
