% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.ced.con}
\alias{f.ced.con}
\title{Determine values for the CED; continuous response}
\usage{
f.ced.con(model.ans, regr.par.matr, track = FALSE)
}
\arguments{
\item{model.ans}{integer, indicates the model that will be fitted; 
one of \code{c(1, 11, 18, 20, 13, 15, 23, 25)}}

\item{regr.par.matr}{numeric matrix, each row contains a vector of parameter
values as needed by the function f.inv.con()}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric vector, calculated CED values
}
\description{
Determine values for the CED; continuous response
}
