% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.ced.pct}
\alias{f.ced.pct}
\title{Calculate CI based on bootstrap estimates and confidence level}
\usage{
f.ced.pct(CED, CED.boot, nruns.intend, sf.x = NA, conf.lev = 0.9,
  track = FALSE)
}
\arguments{
\item{CED}{numeric, the estimated value for CED}

\item{CED.boot}{numeric vector, the estimated CED values following bootstrap}

\item{nruns.intend}{integer, the intended number of bootstrap runs}

\item{sf.x}{numeric, scaling factor for x; default value is NA}

\item{conf.lev}{numeric, the confidence level for confidence intervals; 
default value is 0.9}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, the confidence interval values and the confidence level as 
provided in conf.lev
}
\description{
Calculate CI based on bootstrap estimates and confidence level
}
