% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.cedlines.bin}
\alias{f.cedlines.bin}
\title{Plot dotted lines for the determined CED value for categorical model}
\usage{
f.cedlines.bin(ans.all, inputResponse = NA, inputCED = NA, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{inputResponse}{numeric, value for the response at estimated CED;
default value is NA}

\item{inputCED}{numeric, value for the CED; default value is NA}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\description{
Plot dotted lines for the determined CED value for categorical model
}
