% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.check.cc}
\alias{f.check.cc}
\title{Check whether the parameter value for c is too close to CES}
\usage{
f.check.cc(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
boolean, TRUE if the the parameter value for c is too close to CES; 
if so warning message is printed
}
\description{
Check whether the parameter value for c is too close to CES
}
