% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_data.R
\name{f.check.nonneg.num}
\alias{f.check.nonneg.num}
\title{Check whether the data frame columns are numeric and can be log-transformed}
\usage{
f.check.nonneg.num(dataFrame, track = FALSE)
}
\arguments{
\item{dataFrame}{data frame for which the columns class and values will be checked}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
boolean, TRUE if the vector is not numeric, contains negative values
or contains zero values only; if so warning message is printed
}
\description{
Check whether the data frame columns are numeric and can be log-transformed
}
