% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.CI}
\alias{f.CI}
\title{Wrapper function for calculating CI around parameter estimates, iterates 
f.profile.all()}
\usage{
f.CI(ans.all, nolog = FALSE, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{nolog}{boolean; TRUE if calculations should be performed on the original
scale, FALSE if on the log-scale; default value is FALSE}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Wrapper function for calculating CI around parameter estimates, iterates 
f.profile.all()
}
