% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.control}
\alias{f.control}
\title{Define control parameters for nlminb(), which minimizes the log-likelihood function}
\usage{
f.control(level = 3)
}
\arguments{
\item{level}{integer, one of \code{1:3} determines which control parameters
are used; default value is 3}
}
\value{
list, with control parameters
}
\description{
Define control parameters for nlminb(), which minimizes the log-likelihood function
}
