% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.converged}
\alias{f.converged}
\title{Determine convergence type of the minimization algorithm}
\usage{
f.converged(message, conv.out, track = FALSE)
}
\arguments{
\item{message}{object as returned from nlminb()$message}

\item{conv.out}{object as returned from nlminb()$conv}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
boolean, whether the optimization ended with convergence or not;
with attribute "message" which is the input object message
}
\description{
Determine convergence type of the minimization algorithm
}
