% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.dtype6.mn}
\alias{f.dtype6.mn}
\title{Compare model fitted with and without alfa}
\usage{
f.dtype6.mn(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Compare model fitted with and without alfa
}
