% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.expect.bin}
\alias{f.expect.bin}
\title{Calculate expected response values, for categorical response}
\usage{
f.expect.bin(model.ans = NA, x, regr.par = 0, fct1 = 1, fct2 = 1,
  kk = -1, nn = -1, dtype = 6, CES = NA, ttt = 0, twice = F,
  x.full = NA, fct1.full = NA, fct2.full = NA, ces.ans = 3, CES1, CES2,
  track = FALSE)
}
\arguments{
\item{model.ans}{integer, determines the type of model to be fitted}

\item{x}{numeric vector, the dose values}

\item{regr.par}{numeric vector, regression parameter values}

\item{fct1}{numeric, value for parameter a}

\item{fct2}{numeric, value for parameter b}

\item{kk}{vector of parameters?}

\item{nn}{vector of total number}

\item{dtype}{integer, determines the type of response}

\item{CES}{numeric, value for the CES}

\item{ttt}{numeric, time variable (redundant)}

\item{twice}{logical, if TRUE two parameters are dependent of the same covariate}

\item{x.full}{vector of concentrations}

\item{fct1.full}{numeric, value for parameter a}

\item{fct2.full}{numeric, value for parameter b}

\item{ces.ans}{index type of benchmark response}

\item{CES1}{numeric, value for the first CES (redundant)}

\item{CES2}{numeric, value for the second CES (redundant)}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric vector, the expected response values under the estimated 
regression model
}
\description{
Calculate expected response values, for categorical response
}
