% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.expect.cat}
\alias{f.expect.cat}
\title{computed expected value for categorical model}
\usage{
f.expect.cat(model.type = 2, model.ans, x, regr.par = 0, th.par = 0,
  sig.par = 0, fct1 = 1, fct2 = 1, CES = 0, CES.cat = 1, latent = F,
  ttt = 0, twice, fct3 = 1, ces.ans, decr.zz = T, dtype, fct4 = 1,
  fct5 = 1, fct1.txt = "", fct2.txt = "", fct3.txt = "",
  fct4.txt = "", fct5.txt = "", covar.txt = "", fct3.ref, track = FALSE)
}
\arguments{
\item{model.type}{model type}

\item{model.ans}{type or response fitted}

\item{x}{observations}

\item{regr.par}{parameters}

\item{th.par}{parameter theta}

\item{sig.par}{parameter sigma (standard deviations)?}

\item{fct1}{numeric, value for first parameter}

\item{fct2}{numeric, value for second parameter}

\item{CES}{CES value}

\item{CES.cat}{CES cat}

\item{latent}{latent model}

\item{ttt}{ttt}

\item{twice}{logical, if TRUE two parameters are dependent of the same covariate}

\item{fct3}{numeric, value for third parameter}

\item{ces.ans}{type of benchmark response}

\item{decr.zz}{is decreasing}

\item{dtype}{response data type}

\item{fct4}{numeric, value for fourth parameter}

\item{fct5}{numeric, value for fifth parameter}

\item{fct1.txt}{text for parameter 1}

\item{fct2.txt}{text for parameter 2}

\item{fct3.txt}{text for parameter 3}

\item{fct4.txt}{text for parameter 4}

\item{fct5.txt}{text for parameter 5}

\item{covar.txt}{text for covar}

\item{fct3.ref}{parameter 3 of reference}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
expected value
}
\description{
computed expected value for categorical model
}
