% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.expect.con}
\alias{f.expect.con}
\title{Calculate expected response values, for continuous response}
\usage{
f.expect.con(model.ans, x, regr.par = 0, fct1 = 1, fct2 = 1, fct3 = 1,
  fct4 = 1, fct5 = 1, CES = NA, twice, ttt = 0, y = 0, increase,
  x.mn = NA, par.start = NA, track = FALSE)
}
\arguments{
\item{model.ans}{integer, determines the type of model to be fitted}

\item{x}{numeric vector, the dose values}

\item{regr.par}{numeric vector, regression parameter values}

\item{fct1}{numeric, value for parameter a}

\item{fct2}{numeric, value for parameter b}

\item{fct3}{numeric, value for parameter var}

\item{fct4}{numeric, value for parameter c}

\item{fct5}{numeric, value for parameter d}

\item{CES}{numeric, value for the CES}

\item{twice}{boolean, if some parameter values are equal, see f.execute()}

\item{ttt}{numeric, time variable}

\item{y}{numeric vector, the response values}

\item{increase}{boolean, whether the response values are increasing or 
decreasing for increasing dose values}

\item{x.mn}{numeric value, the mean of dose values, see f.execute()}

\item{par.start}{numeric vector, regression parameter values (e.g. MLEs)}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric vector, the expected response values under the estimated 
regression model
}
\description{
This function is used also for quantal response (f.expect.cat)
}
