% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.hit.constr}
\alias{f.hit.constr}
\title{Check whether any of the MLEs for the parameters hit the lower/upper constraint}
\usage{
f.hit.constr(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Check whether any of the MLEs for the parameters hit the lower/upper constraint
}
