% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.ini}
\alias{f.ini}
\title{Define default parameter values needed in the proast functions}
\usage{
f.ini(odt = NULL, shinyInput, track = FALSE)
}
\arguments{
\item{odt}{list as returned by f.scan()}

\item{shinyInput}{list with values assigned in the shiny app}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
updated version of ans.all, list with all results that were obtained
during the analysis
}
\description{
Define default parameter values needed in the proast functions
}
