% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.inv.bin}
\alias{f.inv.bin}
\title{return the CED value for a set of parameters, CES and ces.ans}
\usage{
f.inv.bin(model.ans, regr.par, CES, ces.ans, track = FALSE)
}
\arguments{
\item{model.ans}{integer, type of response model, only c(3 available}

\item{regr.par}{vector with parameters values (a, b, c)}

\item{CES}{numeric, CES value}

\item{ces.ans}{integer, type of benchmark response}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
CED value
}
\description{
Originally the function can deal all model.ans, 
but because currently this function is only used:
\itemize{
\item{in the \code{f.start.bin} function where model.ans == 16, in which model.ans is set to 3}
\item{in \code{f.ced.cat} for the categorical models: 
	\itemize{
	\item{quantal/binary response: }{(1, 14), 16, 18, 19, 21, 24, 25, 26}
 \item{ordinal response: }{13, 15, 23, 25 so only ces.ans == 1}
 }}
} 
Only this part of the code is retained.
}
