% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.inv.con}
\alias{f.inv.con}
\title{Determine value for the CED; continuous response}
\usage{
f.inv.con(model.ans, params, CES, dtype = 1, track = FALSE)
}
\arguments{
\item{model.ans}{integer, indicates the model that will be fitted; 
one of \code{c(1, 11, 18, 20, 13, 15, 23, 25)}}

\item{params}{numeric vector, parameter values for the model parameters a,
b, c and d}

\item{CES}{value for the CES (by default NULL: not used for continuous models)}

\item{dtype}{response data type, 1 by default}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric, value for the CED
}
\description{
Determine value for the CED; continuous response
}
