% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.lik.cat}
\alias{f.lik.cat}
\title{Calculate likelihood for categorical response}
\usage{
f.lik.cat(theta, x, y = NA, kk, nn, dtype, fct1 = 1, fct2 = 1, nrp,
  nth = NA, nr.aa = 1, nr.bb = 1, model.ans, model.type, CES = NA,
  ttt = 0, twice = NA, alfa.length = 0, x.full = NA, fct1.full = NA,
  fct2.full = NA, ces.ans = 1, fct3 = 1, fct4 = 1, fct5 = 1,
  CES.cat = 1, decr.zz = TRUE, CES1 = CES1, CES2 = CES2, nn.tot, kk.tot,
  xx.tot = xx.tot, fct3.ref, track = FALSE)
}
\arguments{
\item{theta}{numeric vector, the initial regression parameter values}

\item{x}{numeric vector, the dose values}

\item{y}{numeric vector, the response values}

\item{kk}{vector of parameters?}

\item{nn}{vector of total number}

\item{dtype}{integer, determines the type of response}

\item{fct1}{numeric, value for parameter a}

\item{fct2}{numeric, value for parameter b}

\item{nrp}{values?}

\item{nth}{values?}

\item{nr.aa}{values?}

\item{nr.bb}{values?}

\item{model.ans}{integer, type of response model, only 3 available}

\item{model.type}{type of model}

\item{CES}{numeric, value for the CES}

\item{ttt}{numeric, time variable}

\item{twice}{logical, if TRUE two parameters are dependent of the same covariate}

\item{alfa.length}{length of alpha}

\item{x.full}{numeric vector, the dose values}

\item{fct1.full}{numeric, value for parameter a}

\item{fct2.full}{numeric, value for parameter b}

\item{ces.ans}{index type of benchmark response}

\item{fct3}{numeric, value for parameter c}

\item{fct4}{numeric, value for parameter d}

\item{fct5}{numeric, value for parameter e}

\item{CES.cat}{value}

\item{decr.zz}{is z decreasing?}

\item{CES1}{numeric, value for the first CES}

\item{CES2}{numeric, value for the second CES}

\item{nn.tot}{vector of total number}

\item{kk.tot}{vector of parameters?}

\item{xx.tot}{vector of concentrations}

\item{fct3.ref}{reference for parameter 3}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric value, minus the sum of the scores (total likelihood)
}
\description{
Calculate likelihood for categorical response
}
