% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.lik.con}
\alias{f.lik.con}
\title{Calculate likelihood for continuous response}
\usage{
f.lik.con(theta, x, y, dtype, fct1, fct2, fct3, model.ans, mn.log, sd2.log, nn,
  Vdetlim, CES, twice = TRUE, ttt = 0, fct4 = 1, fct5 = 1,
  cens.up = NA, lb = -Inf, ub = Inf, par.tmp, increase = increase,
  x.mn = NA, track = FALSE)
}
\arguments{
\item{theta}{numeric vector, the initial regression parameter values}

\item{x}{numeric vector, the dose values}

\item{y}{numeric vector, the response values}

\item{dtype}{integer, determines the type of response}

\item{fct1}{numeric, value for parameter a}

\item{fct2}{numeric, value for parameter b}

\item{fct3}{numeric, value for parameter var}

\item{model.ans}{integer, determines the model that will be fitted}

\item{mn.log}{numeric vector, transformation of the response values,
see f.execute()}

\item{sd2.log}{numeric vector, transformation of the sd of the response
values, see f.execute()}

\item{nn}{numeric vector, the number of responses per dose level, for 
continuous summary data}

\item{Vdetlim}{numeric vector, values of detection limit}

\item{CES}{numeric, value for the CES}

\item{twice}{boolean, if some parameter values are equal, see f.execute()}

\item{ttt}{numeric, time variable}

\item{fct4}{numeric, value for parameter c}

\item{fct5}{numeric, value for parameter d}

\item{cens.up}{numeric, value for right censoring}

\item{lb}{numeric vector, determines the lower bound for theta;
default value is -Inf}

\item{ub}{numeric vector, determines the upper bound for theta;
default value is Inf}

\item{par.tmp}{numeric vector, regression parameter values, see f.pars()}

\item{increase}{boolean, whether the response values are increasing or 
decreasing for increasing dose values}

\item{x.mn}{numeric value, the mean of dose values, see f.execute()}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric value, minus the sum of the scores (total likelihood)
}
\description{
Calculate likelihood for continuous response
}
