% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.lines.cat}
\alias{f.lines.cat}
\title{plot lines for categorical model}
\usage{
f.lines.cat(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
c(0, y.max)
}
\description{
Note: used in the f.plot.all function, for dtype == 2, 3, 
and where plot.type is set to 3, 5, 9?
so only this part of the code is retained
}
