% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.lines.con}
\alias{f.lines.con}
\title{Plot the curve of the estimated model}
\usage{
f.lines.con(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\description{
Plot the curve of the estimated model
}
