% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.lines.frq}
\alias{f.lines.frq}
\title{Plot the lines}
\usage{
f.lines.frq(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
ans.all$gr.txt
}
\description{
Plot the lines
}
