% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.LL.bin}
\alias{f.LL.bin}
\title{Calculate confidence intervals for each of the plotted points for quantal response}
\usage{
f.LL.bin(k = 4, n = 20, n.pi = 100, conf = 0.9, step = 1.2)
}
\arguments{
\item{k}{integer, number of events in given group}

\item{n}{integer, number of observations in given group}

\item{n.pi}{?}

\item{conf}{?}

\item{step}{?}
}
\value{
vector with lower and upper limit
}
\description{
Calculate confidence intervals for each of the plotted points for quantal response
}
