% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.lump.cat}
\alias{f.lump.cat}
\title{? Add values to e.g. response y to avoid log of 0}
\usage{
f.lump.cat(x, y, nn, fct1, fct2, dtype, twice, track = FALSE)
}
\arguments{
\item{x}{vector independent variable}

\item{y}{vector response}

\item{nn}{vector number of observations}

\item{fct1}{vector covariate on which parameter a is dependent}

\item{fct2}{vector covariate on which parameter b is dependent}

\item{dtype}{integer response data type}

\item{twice}{logical, if TRUE two parameters are dependent of the same covariate}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list with some values
}
\description{
? Add values to e.g. response y to avoid log of 0
}
