% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.mm7.con}
\alias{f.mm7.con}
\title{Main function for bootstrap method for calculating CI of CED}
\usage{
f.mm7.con(ans.all, track = track)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Main function for bootstrap method for calculating CI of CED
}
