% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.model.bb}
\alias{f.model.bb}
\title{change model.ans}
\usage{
f.model.bb(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
updated ans.all
}
\description{
change model.ans
}
