% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.nr.gr}
\alias{f.nr.gr}
\title{Determine the number of group ?}
\usage{
f.nr.gr(nr.aa, nr.bb, twice)
}
\arguments{
\item{nr.aa}{values?}

\item{nr.bb}{values?}

\item{twice}{logical, if TRUE two parameters are dependent of the same covariate}
}
\value{
numeric with bumber of groups?
}
\description{
Determine the number of group ?
}
