% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.P}
\alias{f.P}
\title{Calculate p-value for likelihood test}
\usage{
f.P(loglik.1, loglik.2, df, track = FALSE)
}
\arguments{
\item{loglik.1}{numeric, first likelihood value}

\item{loglik.2}{numeric, second likelihood value}

\item{df}{integer, degrees of freedom}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
numeric, p-value for likelihood test
}
\description{
Calculate p-value for likelihood test
}
