% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.par.u}
\alias{f.par.u}
\title{?}
\usage{
f.par.u(model.ans, regr.par = 0, th.par = 0, nr.aa = 1, nr.bb = 1,
  CES.cat, track = FALSE)
}
\arguments{
\item{model.ans}{type of response}

\item{regr.par}{regression parameter}

\item{th.par}{theta parameter}

\item{nr.aa}{number of a}

\item{nr.bb}{number of b}

\item{CES.cat}{?}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
regr.par.u
}
\description{
called in the f.bb.cat function
}
