% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.pars}
\alias{f.pars}
\title{Construct regression parameter matrix}
\usage{
f.pars(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Construct regression parameter matrix
}
