% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.plot.frq}
\alias{f.plot.frq}
\title{Plot the predicted response values (points) and estimated CIs}
\usage{
f.plot.frq(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
updated ans.all
}
\description{
Plot the predicted response values (points) and estimated CIs
}
