% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.proast}
\alias{f.proast}
\title{Main function to perform proast analysis}
\usage{
f.proast(odt = list(), shinyInput, track = FALSE, debug = FALSE)
}
\arguments{
\item{odt}{list as returned by f.scan()}

\item{shinyInput}{list with values assigned in the shiny app}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}

\item{debug}{logical, if TRUE tryCatch errors and return current object ans.all; default value is FALSE}
}
\value{
ans.all, list with all results that were obtained during the analysis
}
\description{
Main function to perform proast analysis
}
