% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.profile.all}
\alias{f.profile.all}
\title{Profile likelihood method to calculate confidence intervals}
\usage{
f.profile.all(ans.all, nolog = FALSE, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{nolog}{boolean, whether the response is log-transformed or not;
default value is FALSE (ie log-transformed response)}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Profile likelihood method to calculate confidence intervals
}
