% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_data.R
\name{f.remove.NAs}
\alias{f.remove.NAs}
\title{Remove NAs from a given data frame for all listed column indices}
\usage{
f.remove.NAs(variableIndices, originalData, track = FALSE)
}
\arguments{
\item{variableIndices}{numeric vector, column indices for variables to check
for missing values}

\item{originalData}{data frame, data from which missing values should be 
excluded}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
data frame based on original data, all rows with missing values for 
any of the non-zero variables are excluded
}
\description{
Remove NAs from a given data frame for all listed column indices
}
