% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.resid.con}
\alias{f.resid.con}
\title{Calculate residuals: difference of observed and expected response value}
\usage{
f.resid.con(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Calculate residuals: difference of observed and expected response value
}
