% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.split.par}
\alias{f.split.par}
\title{split the vector of parameters}
\usage{
f.split.par(MLE, nrp, nth, dtype, track = FALSE)
}
\arguments{
\item{MLE}{vector of parameters}

\item{nrp}{values?}

\item{nth}{values?}

\item{dtype}{integer, determines the type of response}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list with parameters split by category
}
\description{
split the vector of parameters
}
