% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.start.bin}
\alias{f.start.bin}
\title{Define initial parameter values for the chosen model, binary response}
\usage{
f.start.bin(ans.all, adjust = FALSE, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{adjust}{boolean, if TRUE the parameter start values will be adjusted;
default value is FALSE}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
From f.start.bin of the proast61.3 package, only: tmp.quick == FALSE
}
