% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.start.cat}
\alias{f.start.cat}
\title{Define initial parameter values for the chosen model, categorical response}
\usage{
f.start.cat(ans.all, adjust = FALSE, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{adjust}{boolean, if TRUE the start values of the model parameters will
be adjusted; default value is FALSE}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
from f.start.cat of the proast61.3 package, only: tmp.quick == FALSE
}
