% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.start.con}
\alias{f.start.con}
\title{Define initial parameter values for the chosen model, continuous response}
\usage{
f.start.con(ans.all, adjust = FALSE, fitted = FALSE, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{adjust}{boolean TRUE if start values of the parameters should be adjusted;
default value is FALSE}

\item{fitted}{boolean TRUE if the model has already been fitted;
default value is FALSE}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list, updated version of ans.all
}
\description{
Define initial parameter values for the chosen model, continuous response
}
