% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.start.lm.con}
\alias{f.start.lm.con}
\title{Fit linear model for continuous response}
\usage{
f.start.lm.con(xtmp, ytmp, model.ans, dtype, track = FALSE)
}
\arguments{
\item{xtmp}{numeric vector, values for the independent variable}

\item{ytmp}{numeric vector, values for the dependent variable}

\item{model.ans}{integer, indicates which model will be fitted}

\item{dtype}{integer, indicates the type of response that is used}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
data frame with intercept and slope of the fitted model, the variable
increase is 1 if the slope is positive and -1 otherwise
}
\description{
Fit linear model for continuous response
}
