% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.text.par}
\alias{f.text.par}
\title{Define character vector with all parameter names}
\usage{
f.text.par(ans.all, track = FALSE)
}
\arguments{
\item{ans.all}{list, with all results that were obtained during the analysis}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
character vector, contains all parameter names
}
\description{
Define character vector with all parameter names
}
