% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_visualization.R
\name{f.th.lvm}
\alias{f.th.lvm}
\title{compute th.lvm?}
\usage{
f.th.lvm(th.par, fct3, fct3.ref, track = FALSE)
}
\arguments{
\item{th.par}{parameter theta}

\item{fct3}{initial value for theta}

\item{fct3.ref}{theta reference?}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
th.lvm
}
\description{
compute th.lvm?
}
