% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_continuous.R
\name{f.uniroot.BMDratio}
\alias{f.uniroot.BMDratio}
\title{compute the root of the BMD ratio?}
\usage{
f.uniroot.BMDratio(ratio, cc, CES1, CES2, track = FALSE)
}
\arguments{
\item{ratio}{ratio of interest}

\item{cc}{value for c parameter}

\item{CES1}{first value for CES}

\item{CES2}{second value for CES}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
root
}
\description{
Used in f.expect.con for model.ans == 24 (quantal model)
}
