% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_categorical.R
\name{f.uniroot.probit}
\alias{f.uniroot.probit}
\title{compute the value for the parameter b for a probit function}
\usage{
f.uniroot.probit(aa, CED, CES, ces.ans, track = FALSE)
}
\arguments{
\item{aa}{value for parameter a}

\item{CED}{value for the CES}

\item{CES}{value for the CED}

\item{ces.ans}{index type of benchmark response}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
value for parameter b for probit function
}
\description{
From the original function from proast61.3, only the code for 
model.ans = 11 (original model.ans == 25), used in f.bb.bin is retained.
}
