% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{filterBestCovariates}
\alias{filterBestCovariates}
\title{Retain model with smallest AIC per family when considering covariates}
\usage{
filterBestCovariates(summaryTable)
}
\arguments{
\item{summaryTable}{data frame as returned by the function bindModelResults()}
}
\value{
data frame, subset of summaryTable with one model per family, which 
has the smallest AIC; attribute columnNames same as original summaryTable
}
\description{
Retain model with smallest AIC per family when considering covariates
}
