% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{findBestModel}
\alias{findBestModel}
\title{Find the best model given a table of fitted models results}
\usage{
findBestModel(summaryTable)
}
\arguments{
\item{summaryTable}{data frame as returned by the function bindModelResults()}
}
\value{
if one model in summaryTable NULL, else list with 
\itemize{
	\item{aicWarning: }{character, warning in procedure to find best model}
 \item{acceptedAic: }{boolean, is fitted model accepted based on AIC}
 \item{minBmdl: }{numeric vector, minimum of bmdl values (among accepted models)}
 \item{maxBmdu: }{numeric vector, maximum of bmdu values (among accepted models)}
 \item{minAic: }{numeric, minimum of aic values (among accepted models)}
 \item{bestModel: }{character vector, name and included parameters of best (with minAic) model}
 \item{bestBmd: }{numeric vector, bmd values of best model}
}
}
\description{
Find the best model given a table of fitted models results
}
