% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{findBootstrapBounds}
\alias{findBootstrapBounds}
\title{Find the lower and upper bound for bmd based on parametric bootstrap results}
\usage{
findBootstrapBounds(bootstrapBmd, confidenceLevel = 0.9)
}
\arguments{
\item{bootstrapBmd}{data.frame, bmd values based on parametric bootstrap; 
as returned by bootstrapBmd()}

\item{confidenceLevel}{numeric, defines constructed confidence interval 
for bmd, e.g. if 0.9 then lower and upper bound of 90% CI for bmd are estimated;
default value is 0.9}
}
\value{
numeric vector, estimated lower and upper bound for model-averaged bmd
}
\description{
Find the lower and upper bound for bmd based on parametric bootstrap results
}
