% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{fitAllCovariates}
\alias{fitAllCovariates}
\title{Fit all covariate combinations for model parameters in given model family}
\usage{
fitAllCovariates(data, shinyInput, selectedModel, noFit = FALSE,
  estimateCed = TRUE, track = FALSE)
}
\arguments{
\item{data}{list, as returned from the f.scan() function;}

\item{shinyInput}{list, all parameter values as defined in the shiny UI}

\item{selectedModel, }{integer represents the model to be fit; run function
getModelNames to obtain all models and their coding}

\item{noFit}{boolean, if TRUE the model is not fitted, but default parameter
names and constraints can be obtained; default value is FALSE}

\item{estimateCed}{boolean, if TRUE the ced (i.e. bmd) is estimated along with
confidence interval using profile likelihood; default value is TRUE}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list with savedResults, for each model a list with all results 
obtained during the analysis with fitSingleModel()
}
\description{
Fit all covariate combinations for model parameters in given model family
}
