% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{fitAllModels}
\alias{fitAllModels}
\title{Fit all model families for the continuous or quantal response}
\usage{
fitAllModels(data, shinyInput, fitCovariateCombinations = FALSE,
  noFit = FALSE, estimateCed = TRUE, modelIndices = NULL, track = FALSE)
}
\arguments{
\item{data}{list, as returned from the f.scan() function;}

\item{shinyInput}{list, all parameter values as defined in the shiny UI}

\item{fitCovariateCombinations}{boolean, if TRUE all possible combinations of 
the included covariates for parameters are fitted; default value is FALSE}

\item{noFit}{boolean, if TRUE the model is not fitted, but default parameter
names and constraints can be obtained; default value is FALSE}

\item{estimateCed}{boolean, if TRUE the ced (i.e. bmd) is estimated along with
confidence interval using profile likelihood; default value is TRUE}

\item{modelIndices}{numeric vector, indices of the models to be fit; if NULL
values are obtained from getModelNames(); default value is NULL}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list with for each model a list with all results obtained during 
the analysis with fitSingleModel()
}
\description{
Fit all model families for the continuous or quantal response
}
