% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{getModelNames}
\alias{getModelNames}
\title{Index in Proast and full names of the fitted models}
\usage{
getModelNames(dtype)
}
\arguments{
\item{dtype}{integer, indicating the response type}
}
\value{
named vector with model indices for specific data type and names the 
model names
}
\description{
Index in Proast and full names of the fitted models
}
