% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{getModelParameters}
\alias{getModelParameters}
\title{Obtain the available model parameters for a given model family}
\usage{
getModelParameters(selectedModel = NA, dtype)
}
\arguments{
\item{selectedModel, }{integer represents the model to be fit; run function
getModelNames to obtain all models and their coding; default value is NA}

\item{dtype}{integer, indicating the response type}
}
\value{
vector with names the model parameters
}
\description{
Obtain the available model parameters for a given model family
}
