% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{makeForestPlot}
\alias{makeForestPlot}
\title{Forest plot for estimated bmd, bmdl and bmdu values per response and covariate}
\usage{
makeForestPlot(bmdData, groupNames = NULL)
}
\arguments{
\item{bmdData}{data.frame with response name, estimated bmd, bmdl and bmdu 
values}

\item{groupNames}{character vector, names of the covariate groups;
default value is NULL}
}
\value{
no object is returned, show plot in current device
}
\description{
Forest plot for estimated bmd, bmdl and bmdu values per response and covariate
}
