% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{matchWithResults}
\alias{matchWithResults}
\title{Find which fitted models' results match the listed model and parameter names}
\usage{
matchWithResults(savedResults, modelNames, parameterNames = NULL)
}
\arguments{
\item{savedResults}{list as returned by the function fitAllModels() or 
fitAllCovariates()}

\item{modelNames}{character vector, names of the models to be selected; 
see getModelNames() for available options}

\item{parameterNames}{character vector, same length as \code{modelNames},
names of the model parameters to be selected; see getModelParameters() for
available options; default value is NULL}
}
\value{
numeric vector, indices of the matching fitted results in the 
list \code{savedResults}
}
\description{
Find which fitted models' results match the listed model and parameter names
}
