% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{optimizeBmd}
\alias{optimizeBmd}
\title{Estimate the model-averaged BMD using numeric optimization techniques}
\usage{
optimizeBmd(weights, modelResults, nIterations = 400, naiveApproach = FALSE)
}
\arguments{
\item{weights}{numeric vector, estimated weights as returned by calculateWeights()}

\item{modelResults}{list, with results for each model, same length as weights. 
For each model a list with at least npar, loglik, x, y, model.ans, regr.par, 
CES and ces.ans; these are by default included in result from f.proast(). 
Eventually contains also fct1 and fct2 if factors are included for the model parameters.}

\item{nIterations}{integer, number of iterations for numeric optimization;
default value is 400}

\item{naiveApproach}{boolean, TRUE if the model-averaged BMD is estimated as the 
weighted average of bmd values, FALSE if the model-averaged BMD is estimated
based on weighted average of response values; default value is FALSE}
}
\value{
numeric, the model-averaged BMD; error is returned if the numeric 
procedure did not converge after \code{nIterations} iterations
}
\description{
Estimate the model-averaged BMD using numeric optimization techniques
}
