% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{pasteToVector}
\alias{pasteToVector}
\title{Paste elements of vector into string vector (for testing)}
\usage{
pasteToVector(x)
}
\arguments{
\item{x}{vector}
}
\value{
string of the form "c(<elements of x>)"
}
\description{
Paste elements of vector into string vector (for testing)
}
