% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{plotAverageModel}
\alias{plotAverageModel}
\title{Plot for the model-averaged response values and bmd}
\usage{
plotAverageModel(proastData, xans, yans, nans = 0, bmd, modelResults,
  bootstrapBmd = NULL, bootstrapModelResults = NULL,
  confidenceLevel = 0.9, naiveApproach = FALSE)
}
\arguments{
\item{proastData}{list, data in proast format as returned by f.scan()}

\item{xans}{integer, indicates the column in proastData$data
containing x-values (dose)}

\item{yans}{integer, indicates the column in proastData$data 
containing y-values (response)}

\item{nans}{integer, indicates the column in proastData$data 
containing sample size; if irrelevant set to 0; default value is 0}

\item{bmd}{numeric value, the estimated model-averaged BMD as returned by 
optimizeBmd()}

\item{modelResults}{list, with results for fitted model for bmd; contains 
at least npar, loglik, model.ans, regr.par, CES and ces.ans; 
these are by default included in result from f.proast(); Eventually contains 
also fct1 and fct2 if factors are included for the model parameters}

\item{bootstrapBmd}{numeric vector, estimated bmd for each of the bootstrap data 
sets as returned by bootstrapBmd(); default value is NULL}

\item{bootstrapModelResults}{list with for each of the bootstrap data sets 
results of the fitted model; should contain at least see param modelResults above;
as returned by bootstrapBmd(); default value is NULL}

\item{confidenceLevel}{numeric, defines constructed confidence interval 
for bmd, e.g. if 0.9 then lower and upper bound of 90% CI for bmd are estimated;
default value is 0.9}

\item{naiveApproach}{boolean, TRUE if the model-averaged BMD is estimated as the 
weighted average of bmd values, FALSE if the model-averaged BMD is estimated
based on weighted average of response values; default value is FALSE}
}
\value{
no return value; plot is written to the current device
}
\description{
Plot for the model-averaged response values and bmd
}
