% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{printer}
\alias{printer}
\title{Print for debugging}
\usage{
printer(x)
}
\arguments{
\item{x}{R object that will be printed}
}
\value{
NULL, print output in the console
}
\description{
Print for debugging
}
