% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{summaryModels}
\alias{summaryModels}
\title{Summarize the results of all fitted models for continuous or quantal response}
\usage{
summaryModels(savedResults)
}
\arguments{
\item{savedResults}{list as returned by the function fitAllModels() or 
fitAllCovariates()}
}
\value{
list with summaryTable and extraInfo, respectively as returned by 
bindModelResults() and findBestModel()
}
\description{
Summarize the results of all fitted models for continuous or quantal response
}
