% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testConstraints}
\alias{testConstraints}
\title{Test function for defining parameter constraints}
\usage{
testConstraints(shinyInput, data, modelAns, doPrint, track)
}
\arguments{
\item{shinyInput}{list with input from Shiny}

\item{data}{dataset}

\item{modelAns}{integer, type of response fitted}

\item{doPrint}{logical, if TRUE, additional information concerning the test is printed}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
no returned value
}
\description{
Test function for defining parameter constraints
}
