% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testCovariates}
\alias{testCovariates}
\title{Test function for including covariates}
\usage{
testCovariates(shinyInput, shinyInputCovariates, data, modelAns,
  results = NULL, doPrint, track)
}
\arguments{
\item{shinyInput}{list with input from Shiny}

\item{shinyInputCovariates}{list with input from Shiny for parameters
concerning the covariates}

\item{data}{dataset}

\item{modelAns}{integer, type of response fitted}

\item{results}{list with expected results, should contain the following items
\itemize{
    \item{par.start: }{list with starting values for the parameters, of length \code{modelAns}}
    \item{loglik: }{vector with values for the log likelihood, of length \code{modelAns}}
    \item{CED: }{vector with CED values, of length \code{modelAns}}
    \item{conf.int: }{vector with confidence intervals, with 2 rows (lower/upper bounds)
    and length \code{modelAns} columns}
}}

\item{doPrint}{logical, if TRUE, additional information concerning the test is printed}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
no returned value
}
\description{
Test function for including covariates
}
