% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testStartingValues}
\alias{testStartingValues}
\title{Test function for defining starting values}
\usage{
testStartingValues(shinyInput, data, modelAns, startValues = NULL,
  results = NULL, doPrint, track)
}
\arguments{
\item{shinyInput}{list with input from Shiny}

\item{data}{dataset}

\item{modelAns}{integer, type of response fitted}

\item{startValues}{list with starting values for each of the response fitted
This should be named with the modelAns as character; default value is NULL}

\item{results}{list with expected results, default value is NULL;
should contain the following items
\itemize{
    \item{par.start: }{list with starting values for the parameters, of length \code{modelAns}}
    \item{loglik: }{vector with values for the log likelihood, of length \code{modelAns}}
    \item{CED: }{(only for continous model) vector with CED values, of length \code{modelAns}}
    \item{CED.matr: }{(only for quantal model) vector with CED values, 
of length (\code{modelAns} - 2) (without null and full models)}
    \item{conf.int: }{vector with confidence intervals, with 2 rows (lower/upper bounds)
    and with number of columns, for:
    \itemize{
        \item{continuous model}{length \code{modelAns}}
        \item{quantal model}{length (\code{modelAns} - 2) (without null and full models)}
    }}
}}

\item{doPrint}{logical, if TRUE, additional information concerning the test is printed}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
no returned value
}
\description{
Test function for defining starting values
}
